認証機能修正
